Made a better attempt at getting sorting working, even though it still doesn't
