+++ src/app/user-articles/article-list/article-list.component.html
@@ -8,7 +8,7 @@
   <!-- Description Column -->
   <ng-container matColumnDef="description">
     <mat-header-cell *matHeaderCellDef mat-sort-header> Description </mat-header-cell>
-    <mat-cell *matCellDef="let row"> {{row.description}}% </mat-cell>
+    <mat-cell *matCellDef="let row"> {{row.description}} </mat-cell>
   </ng-container>
 
   <!-- Date Posted Column -->
